Mark resources with our config name so we can delete them
In most cases, when a resource is deleted and we need to clean up its
EPIC-side counterpart, we can follow the resource "chain" up through
the Gateway and the GatewayClass to the GatewayClassConfig to get the
data that we need to connect to EPIC. In some cases, though, this
isn't possible. If the user deletes a Gateway, for example, we want to
be able to clean up the HTTPRoutes that reference it.

Now when we announce a resource to EPIC we add an annotation that
indicates which GatewayConfigClass was used to create the EPIC-side
resource (in addition to the link that we were previously
adding). This gives us enough data to delete resources on EPIC even if
the chain is broken.