Add TrueIngress components to the agent
TrueIngress cli executables are installed to /opt/acnodal/bin in the
agent container.

We run the pinger in a separate image in the agent pod instead of
launching it explicitly using go code. This seems more k8s-ish.

The TrueIngress setup code in internal/trueingress was copied from
PureLB internal/pfc.