Run TrueIngress pinger in the agent container instead of a sidecar
For reasons that I don't have time to figure out now, if I run the
pinger in its own container it doesn't see the TrueIngress maps loaded
in the agent container. So let's run the pinger like we used to, i.e.,
as a background process in the agent container.