Add EndpointSliceShadow because we can't add annotations to slices
For some fucked-up reason the people who implemented the EndpointSlice
controller chose to trash annotations and finalizers when the service
scales down to zero endpoints:
https://github.com/kubernetes/kubernetes/issues/108267

Since we can't use EndpointSlice in the way in which it was intended,
we have to implement a "shadow" resource. Basically, for every ES that
we care about, we also create an EndpointSliceShadow to store the
stuff that we would have stored in annotations, which is basically the
info that we need to clean up the object if it is deleted.

The shadow is also the flag that indicates when an ES has been
announced to EPIC. If there's a shadow then the ES reconciler will
send any updates to EPIC.

The HTTPRoutes controller announces slices to EPIC because it has all
the info that it needs to do so. Once a slice has been announced, the
ES controller can process updates and cleanup, but the first
announcement needs info in the HTTPRoute.