Add a second deployment, service, and route backend ref to samples
This demonstrates traffic splitting within a route rule. We've got two
services and the rule splits the traffic two-to-one between the
services.

You can tell which cluster you get for a given request using the
Echoserver "Server Cluster" field. It will alternate between
"epic-sample" and "epic-sample2" with twice as many requests going to
"epic-sample".