Add a "sharing key" annotation to the Gateway
By virtue of our distributed architecture, we should be able to have
xRoutes on multiple client clusters share a single Gateway, since that
Gateway runs on EPIC. This is the first step in that direction. If the
user annotates a Gateway with the acnodal.io/epic-sharing-key then the
EPIC client will attempt to rendez-vous with that Gateway instead of
creating a new one.

The sharing key is the UID of the EPIC GWProxy resource. We need to
make it more user-visible.