Nudge a Gateway's child routes when the Gateway changes
This handles the case where a Gateway is deleted and re-added. In that
case the HTTPRoutes need to update their announcements to match the
new Gateway's UID. The new Gateway has the same name as the old one
but it's actually a different k8s resource so the EPIC resources need
to be updated to match.