Treat no namespace as "default" in GatewayClassConfig ref
This is a case where the spec is pretty user-hostile so I'm going to
go a more user-friendly route. The spec[1] requires a namespace field
if the GatewayClassConfig is namespaced, which ours is. So in theory
we should fail if a GatewayClass doesn't have a namespace in the ref
to GatewayClassConfig. I think it's better to treat no namespace as
"default" so a user can drop a GatewayClassConfig into the default NS
and refer to it without having to add an explicit namespace.