Attach encapper to all ENI interfaces
Amazon allocates more than one ENI to each node and their CNI sets up
routes so that each pod's traffic is routed through one of the
ENI's. This broke the PFC since the encapper was only attached to one
of the ENIs.

Now if we think that the node is an EKS node we attach the encapper to
all of the network interfaces that match "eth[1-9]" in addition to the
default interface, which is usually eth0 in this case.

We use Node.Spec.ProviderID to decide when we're running in an EKS
environment. It's a URI-encoded string whose protocol is "aws" when
running in an EKS environment and it's "" when running in an ad-hoc
environment.