Implicitly enable packet forwarding on EKS secondary ENIs
EKS allocates additional ENIs based on the number of IP addresses
that it needs to support on each node. We need to forward packets
after doing the FIB lookup on the secondary ENIs (i.e., eth1-9) but we
don't want to make a big change like enabling forwarding everywhere
so for now we're going to implicitly add the forward flag where we
think we need it.