Don't use the downward API to find the node address
We need node addresses in two places: the manager needs them to help
EPIC build the TrueIngress maps, and the per-node agent needs them to
find their config from the set of node configs. We used to use the
downward API in the agent pods because each agent only needs its own
address.

Something changed in k8s v1.26, though, so this will no longer
work. The host IP returned from the downward API used to match the
first internal address in the node resource, but it no longer
does. This caused the agents to get confused because they couldn't
find their configurations in the maps.

Now we no longer use the downward API to get the node addresses. In
its place there's now a method that lets you specify an address
family. We need this because TrueIngress doesn't yet work with IPV6.

Signed-off-by: Toby Cabot <toby@acnodal.io>