Run build-time tools using "go run" instead of installing them
The default for the upstream project is to install build-time tools
like controller-gen and kustomize into a local bin folder and run them
from there. This requires some weird makefile code and is likely a
holdover from the days before "go run" could work in a module-aware
way.

Now we use "go run" to run these tools. We run them with a version
number so they'll use their own go.mod instead of ours which ensures
that their dependencies are what they expect.