Reduce GUE ping interval to 3s
When we first coded this we set the value to to the longest time that
we thought would work, in order to be as efficient as possible with
the network. My time is more valuable than a few network packets,
though, and this change cuts 2 minutes off the time it takes to run
the Gateway API conformance test.