Don't duplicate Parent status updates
I can't figure out what the GatewayRef member of the RouteStatusUpdate
struct does. A route can reference multiple parents so it's not clear
why we have a field that can hold only one reference.

I was having a problem where each time the controller fired, it
duplicated all of the parent statuses. This was caused by this
code. If I figure out what it was doing then I might need to put it
back but for now it works better without it.